Set shell script interpreter to bash instead of sh
On many systems sh points to bash. In that case the scripts work fine. However, double square brackets won't necessarily work when sh does not point to bash but to a POSIX sh. This commit changes the interpreter to bash. Alternatively, one can change the double square brackets to single square brackets to achieve portability.